requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=9.5.0
easyocr>=1.7.0
pytesseract>=0.3.10

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Time Series Forecasting
prophet>=1.1.4
statsmodels>=0.14.0

# Backend API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Frontend
streamlit>=1.28.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6

# Database
sqlite3  # Built-in with Python

# Notifications and Communication
twilio>=8.5.0
python-telegram-bot>=20.3
smtplib  # Built-in with Python
email-validator>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.1.0
python-jose[cryptography]>=3.3.0
python-dateutil>=2.8.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
pre-commit>=3.3.0

# Deployment
gunicorn>=21.2.0
docker-compose>=1.29.0

# Image Processing
imageio>=2.31.0
albumentations>=1.3.0